Extract AsyncPoll class for future use in dynamic gas price service
